feat: unified multiline_input that opens at cursor, adapts to window width
